src/familly_nexus/backend/requirements.txt
Flask
flask_sqlalchemy
flask_cors
psycopg2-binary
python-dotenv
gunicorn
requests
Pillow  # For image processing (extracting EXIF data)
